+++ airBnbApp/src/main/java/com/codingakash/projects/airBnbApp/service/RoomServiceImpl.java
@@ -4,7 +4,9 @@ import com.codingakash.projects.airBnbApp.dto.HotelDto;
 import com.codingakash.projects.airBnbApp.dto.RoomDto;
 import com.codingakash.projects.airBnbApp.entity.Hotel;
 import com.codingakash.projects.airBnbApp.entity.Room;
+import com.codingakash.projects.airBnbApp.entity.User;
 import com.codingakash.projects.airBnbApp.exception.ResourceNotFoundException;
+import com.codingakash.projects.airBnbApp.exception.UnAuthorizedException;
 import com.codingakash.projects.airBnbApp.repository.HotelRepository;
 import com.codingakash.projects.airBnbApp.repository.InventoryRepository;
 import com.codingakash.projects.airBnbApp.repository.RoomRepository;
@@ -12,6 +14,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -35,6 +38,13 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel Not found with this id  " + hotelId));
+
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("this User does not own this hotel with hotelId "+hotelId);
+        }
+
         Room room = modelMapper.map(roomDto , Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -55,6 +65,11 @@ if(hotel.getActive()){
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel Not found with this id  " + hotelId));
 
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("this User does not own this hotel with hotelId "+hotelId);
+        }
 
         return hotel.getRooms()
                 .stream()
@@ -81,6 +96,12 @@ if(hotel.getActive()){
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(()->new ResourceNotFoundException("Room not found with this id :" + roomId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorizedException("this User does not own this room with id "+roomId);
+        }
 
         inventoryService.deleteAllInventories(room);
